Making configuratio path passing optional
Instead of passing the path to activity configuration upon each request
the runner has a default path to search activity configuration files
from. The default path points to `app/courses`. The best way would be to
simply symlink the courses repo to that location and that's it. However,
the configuration path can still be overridden by passing it in each
request.