Fixing a bug with multiple files in the RunCommand
From looking at the KnpU repo, I can see that multiple files should be separated by the 003 char.
BUT, I didn't see that represented anywhere in this command.

The fix is a bit ugly, but for multiple files and stdin, I needed to read the stream all the way through
and collect and assign the content to different files as I saw the 003 character.