Fixing and clarifying a problem with resolved and non-resolved paths
Primarily, the "skeletons" key on Activity is almost not needed. We don't actually read the skeleton files (because
the user is submitting his/her answers for the skeletons) so we only have them in Activity so that we can have some
extra validation that input files and contexts are included in the allowed set of files.

Related to this, almost none of the paths on Activity need to be resolved to their absolute location. Most (i.e. the
input files) just contain the actual, submitted contents. There are just a few things that are not submitted (context,
suite) where we need to have the absolute path so we can go find these files. This removes the resolving of the path
for skeleton files and for the entry point.