Working to simplify the testing process
This introduces an ActivitySuiteInterface, with a runTest method, which is passed
an object. The idea is that we could hopefully have some real phpunit tests in the
suite to test that different variations are working/not working. This makes the
suite class a lot more re-usable. We haven't fully realized this yet, but that's the plan.