Fixing a subtle bug when we want to add a file (not its contents directly) to the FileBuilder
When creating an activity, you might want to point to an existing file, so you use
something like $builder->addFile('Product.php', __DIR__.'/Product.php'); But, that
absolute file path will not be accessible when grading. Fortunately, we don't need
to read __DIR__.'/Product.php' at that point, because the `Product.php` source is
sent to the API. This avoids trying the file_get_contents() unless we absolutely
need it.